move Stream imp to separated mod file #218
